Migrates ClientImagesList to BaseDreamList
