+++ OpenDreamRuntime/Objects/Types/Lists/DreamList.cs
@@ -108,29 +108,6 @@ public class DreamList : BaseDreamList {
         base.HandleDeletion();
     }
 
-    public override BaseDreamList CreateCopy(int start = 1, int end = 0) {
-        if (start == 0) ++start; //start being 0 and start being 1 are equivalent
-
-        var values = GetValues();
-        if (end > values.Count + 1 || start > values.Count + 1) throw new Exception("list index out of bounds");
-        if (end == 0) end = values.Count + 1;
-        if (end <= start)
-            return new DreamList(ObjectDefinition, 0);
-
-        List<DreamValue> copyValues = values.GetRange(start - 1, end - start);
-
-        Dictionary<DreamValue, DreamValue>? associativeValues = null;
-        if (_associativeValues != null) {
-            associativeValues = new(end - start);
-            foreach (var key in copyValues) {
-                if (_associativeValues.TryGetValue(key, out var value))
-                    associativeValues[key] = value;
-            }
-        }
-
-        return new DreamList(ObjectDefinition, copyValues, associativeValues);
-    }
-
     /// <summary>
     /// Returns the list of array values. Doesn't include the associative values indexable by some of these.
     /// </summary>
@@ -1084,71 +1061,6 @@ public sealed class ClientScreenList(DreamObjectTree objectTree, ServerScreenOve
     public override int FindValue(DreamValue value, int start = 1, int end = 0) {
         throw new NotImplementedException($".Find() is not yet implemented on {GetType()}");
     }
-}
-
-// client.images list
-public sealed class ClientImagesList(DreamObjectTree objectTree, ServerClientImagesSystem? clientImagesSystem, DreamConnection connection) : DreamList(objectTree.List.ObjectDefinition, 0) {
-    private readonly List<DreamValue> _imageObjects = new();
-
-    public override DreamValue GetValue(DreamValue key) {
-        if (!key.TryGetValueAsInteger(out var imageIndex) || imageIndex < 1 || imageIndex > _imageObjects.Count)
-            throw new Exception($"Invalid index into client images list: {key}");
-
-        var value = _imageObjects[imageIndex - 1];
-        value.IncRef();
-        return value;
-    }
-
-    public override List<DreamValue> GetValues() {
-        return _imageObjects;
-    }
-
-    public override IEnumerable<DreamValue> EnumerateValues() {
-        return _imageObjects;
-    }
-
-    public override void SetValue(DreamValue key, DreamValue value, bool allowGrowth = false) {
-        throw new Exception("Cannot write to an index of a client images list");
-    }
-
-    public override void AddValue(DreamValue value) {
-        if (!value.TryGetValueAsDreamObject<DreamObjectImage>(out var image))
-            return;
-
-        clientImagesSystem?.AddImageObject(connection, image);
-        _imageObjects.Add(value);
-    }
-
-    public override void RemoveValue(DreamValue value) {
-        if (!value.TryGetValueAsDreamObject<DreamObjectImage>(out var image))
-            return;
-
-        clientImagesSystem?.RemoveImageObject(connection, image);
-        _imageObjects.Remove(value);
-    }
-
-    public override void Cut(int start = 1, int end = 0) {
-        if (end == 0 || end > _imageObjects.Count + 1) end = _imageObjects.Count + 1;
-
-        for (int i = start - 1; i < end - 1; i++) {
-            if (!_imageObjects[i].TryGetValueAsDreamObject<DreamObjectImage>(out var image))
-                continue;
-
-            clientImagesSystem?.RemoveImageObject(connection, image);
-        }
-
-        _imageObjects.RemoveRange(start - 1, end - start);
-    }
-
-    public override int GetLength() {
-        return _imageObjects.Count;
-    }
-
-    public override int FindValue(DreamValue value, int start = 1, int end = 0) {
-        throw new NotImplementedException($".Find() is not yet implemented on {GetType()}");
-    }
-
-    public override bool ContainsValue(DreamValue value) => _imageObjects.Contains(value);
 }
 
 // world.contents list
